rustc_target: move `LayoutOf` to `ty::layout`.
