+++ src/tools/clippy/clippy_lints/src/large_stack_arrays.rs
@@ -4,10 +4,9 @@ use if_chain::if_chain;
 use rustc_hir::{Expr, ExprKind};
 use rustc_lint::{LateContext, LateLintPass};
 use rustc_middle::mir::interpret::ConstValue;
+use rustc_middle::ty::layout::LayoutOf;
 use rustc_middle::ty::{self, ConstKind};
 use rustc_session::{declare_tool_lint, impl_lint_pass};
-
-use crate::rustc_target::abi::LayoutOf;
 
 declare_clippy_lint! {
     /// ### What it does
